pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ralph-agent"
version = "0.1.0"
description = "An autonomous coding agent that runs in a loop"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Your Name", email = "you@example.com" },
]
keywords = [
    "ai",
    "agent",
    "claude",
    "coding",
    "autonomous",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "claude-code-sdk>=0.0.25",
    "rich>=13.0",
    "python-telegram-bot>=21.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
ralph = "ralph.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["ralph"]
